Create README.md
I am in the process of completing my final year as an Information Technology undergraduate student at York University. I am a quick and eager learner who is seeking a position related to Web Development, Database Administration, or IT Security with a progressive and trustworthy organization. My proficient communication and conflict resolution skills, along with experience in a variety of programming languages that include but are not limited to Java, SQL, and JavaScript, provide me with the ability to deliver innovative ideas that assist with the technological development of the organization. I aim to make a positive impact within any organization I join, through teamwork, communication, problem-solving, and provision of new and innovative ideas.